requirements.txt
fastapi
pydantic
transformers
huggingface_hub
accelerate
torch
fastapi-cache2
redis
uvicorn
python-dotenv
bitsandbytes